requirements.txt
Flask>=3.0.1
Werkzeug>=3.0.1
PyPDF2>=3.0.1
openai==0.28
python-dotenv